requirements.txt
roboflow
ultralytics
matplotlib
huggingface-hub
albumentations
ipython
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics-thop
urllib3==2.2.3
uvicorn==0.32.1
zipp==3.21.0
